+++ cargo-nextest/src/helpers.rs
@@ -1,0 +1,21 @@
+// Copyright (c) The nextest Contributors
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+#[cfg(feature = "self-update")]
+pub(crate) fn log_needs_update(level: log::Level, extra: &str) {
+    use owo_colors::{OwoColorize, Stream};
+
+    log::log!(
+        level,
+        "update nextest with {}{}",
+        "cargo nextest self update".if_supports_color(Stream::Stderr, |x| x.bold()),
+        extra,
+    );
+}
+
+#[cfg(not(feature = "self-update"))]
+pub(crate) fn log_needs_update(level: log::Level, extra: &str) {
+    log::log!(level, "update nextest via your package manager{}", extra);
+}
+
+pub(crate) const BYPASS_VERSION_TEXT: &str = ", or bypass check with --override-version-check";