+++ nextest-runner/src/config/config_impl.rs
@@ -3,8 +3,8 @@
 
 use super::{
     CompiledOverride, CompiledOverridesByProfile, CustomTestGroup, DeserializedOverride,
-    RetryPolicy, SettingSource, SlowTimeout, TestGroup, TestGroupConfig, TestSettings, TestThreads,
-    ThreadsRequired, ToolConfigFile,
+    NextestVersionConfig, NextestVersionDeserialize, RetryPolicy, SettingSource, SlowTimeout,
+    TestGroup, TestGroupConfig, TestSettings, TestThreads, ThreadsRequired, ToolConfigFile,
 };
 use crate::{
     errors::{
@@ -51,6 +51,7 @@ pub fn get_num_cpus() -> usize {
 pub struct NextestConfig {
     workspace_root: Utf8PathBuf,
     inner: NextestConfigImpl,
+    nextest_version: NextestVersionConfig,
     overrides: CompiledOverridesByProfile,
 }
 
@@ -135,7 +136,7 @@ impl NextestConfig {
     {
         let workspace_root = workspace_root.into();
         let tool_config_files_rev = tool_config_files.into_iter().rev();
-        let (inner, overrides) = Self::read_from_sources(
+        let (inner, nextest_version, overrides) = Self::read_from_sources(
             graph,
             &workspace_root,
             config_file,
@@ -145,6 +146,7 @@ impl NextestConfig {
         Ok(Self {
             workspace_root,
             inner,
+            nextest_version,
             overrides,
         })
     }
@@ -177,9 +179,15 @@ impl NextestConfig {
         Self {
             workspace_root: workspace_root.into(),
             inner: deserialized.into_config_impl(),
+            nextest_version: NextestVersionConfig::default(),
             // The default config does not (cannot) have overrides.
             overrides: CompiledOverridesByProfile::default(),
         }
+    }
+
+    /// Returns the nextest version requirement.
+    pub fn nextest_version(&self) -> &NextestVersionConfig {
+        &self.nextest_version
     }
 
     /// Returns the profile with the given name, or an error if a profile was specified but not
@@ -201,13 +209,24 @@ impl NextestConfig {
         file: Option<&Utf8Path>,
         tool_config_files_rev: impl Iterator<Item = &'a ToolConfigFile>,
         unknown_callback: &mut impl FnMut(&Utf8Path, Option<&str>, &BTreeSet<String>),
-    ) -> Result<(NextestConfigImpl, CompiledOverridesByProfile), ConfigParseError> {
+    ) -> Result<
+        (
+            NextestConfigImpl,
+            NextestVersionConfig,
+            CompiledOverridesByProfile,
+        ),
+        ConfigParseError,
+    > {
         // First, get the default config.
         let mut composite_builder = Self::make_default_config();
 
         // Overrides are handled additively.
         // Note that they're stored in reverse order here, and are flipped over at the end.
         let mut overrides = CompiledOverridesByProfile::default();
+
+        // The nextest version configuration is handled specially: highest version specified by any
+        // tool wins.
+        let mut nextest_version = NextestVersionConfig::default();
 
         let mut known_groups = BTreeSet::new();
 
@@ -222,6 +241,7 @@ impl NextestConfig {
                 source.clone(),
                 &mut overrides,
                 unknown_callback,
+                &mut nextest_version,
                 &mut known_groups,
             )?;
 
@@ -247,6 +267,7 @@ impl NextestConfig {
             source.clone(),
             &mut overrides,
             unknown_callback,
+            &mut nextest_version,
             &mut known_groups,
         )?;
 
@@ -263,7 +284,7 @@ impl NextestConfig {
             override_.reverse();
         }
 
-        Ok((config.into_config_impl(), overrides))
+        Ok((config.into_config_impl(), nextest_version, overrides))
     }
 
     #[allow(clippy::too_many_arguments)]
@@ -275,6 +296,7 @@ impl NextestConfig {
         source: File<FileSourceFile, FileFormat>,
         overrides_out: &mut CompiledOverridesByProfile,
         unknown_callback: &mut impl FnMut(&Utf8Path, Option<&str>, &BTreeSet<String>),
+        nextest_version: &mut NextestVersionConfig,
         known_groups: &mut BTreeSet<CustomTestGroup>,
     ) -> Result<(), ConfigParseError> {
         // Try building default builder + this file to get good error attribution and handle
@@ -286,6 +308,10 @@ impl NextestConfig {
 
         if !unknown.is_empty() {
             unknown_callback(config_file, tool, &unknown);
+        }
+
+        if let Some(version) = this_config.nextest_version.clone() {
+            nextest_version.accumulate(version, tool);
         }
 
         // Check that test groups are named as expected.
@@ -700,6 +726,8 @@ impl NextestConfigImpl {
 #[serde(rename_all = "kebab-case")]
 struct NextestConfigDeserialize {
     store: StoreConfigImpl,
+    #[serde(default)]
+    nextest_version: Option<NextestVersionDeserialize>,
     #[serde(default)]
     test_groups: BTreeMap<CustomTestGroup, TestGroupConfig>,
     #[serde(rename = "profile")]